# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ContainerScope API
  description: Multi-tenant DevOps observability and security visualizer API
  version: 0.1.0
  contact:
    name: ContainerScope Support
    email: support@containerscope.io
  license:
    name: MIT

servers:
  - url: http://localhost:8080
    description: Development
  - url: https://api.containerscope.io
    description: Production

tags:
  - name: Auth
    description: Authentication endpoints
  - name: Organizations
    description: Organization management
  - name: Connections
    description: Agent connections
  - name: Topology
    description: Container topology
  - name: Metrics
    description: Resource metrics
  - name: Flows
    description: Network flows
  - name: Vulnerabilities
    description: Security scanning
  - name: Alerts
    description: Alerting
  - name: System
    description: System endpoints

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key

  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        code:
          type: integer
        message:
          type: string

    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        created_at:
          type: string
          format: date-time

    Org:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        plan:
          type: string
        created_at:
          type: string
          format: date-time

    Connection:
      type: object
      properties:
        id:
          type: string
          format: uuid
        org_id:
          type: string
          format: uuid
        type:
          type: string
          enum: [docker, k8s]
        name:
          type: string
        status:
          type: string
        agent_token:
          type: string
        last_seen_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time

    Container:
      type: object
      properties:
        id:
          type: string
          format: uuid
        connection_id:
          type: string
          format: uuid
        runtime_id:
          type: string
        name:
          type: string
        image:
          type: string
        state:
          type: string
        labels:
          type: object
        ports:
          type: array
          items:
            $ref: '#/components/schemas/Port'
        created_at:
          type: string
          format: date-time

    Port:
      type: object
      properties:
        host_port:
          type: string
        container_port:
          type: string
        protocol:
          type: string

    Network:
      type: object
      properties:
        id:
          type: string
          format: uuid
        connection_id:
          type: string
          format: uuid
        name:
          type: string
        driver:
          type: string
        scope:
          type: string
        subnet:
          type: string

    Edge:
      type: object
      properties:
        id:
          type: string
          format: uuid
        connection_id:
          type: string
          format: uuid
        src_container_id:
          type: string
          format: uuid
        dst_container_id:
          type: string
          format: uuid
        dst_ip:
          type: string
        dst_port:
          type: integer
        protocol:
          type: string

    Topology:
      type: object
      properties:
        containers:
          type: array
          items:
            $ref: '#/components/schemas/Container'
        networks:
          type: array
          items:
            $ref: '#/components/schemas/Network'
        edges:
          type: array
          items:
            $ref: '#/components/schemas/Edge'

    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string

    Membership:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
        org_id:
          type: string
          format: uuid
        role:
          type: string
          enum: [owner, admin, member, viewer]
        created_at:
          type: string
          format: date-time

paths:
  /healthz:
    get:
      summary: Health check
      tags: [System]
      responses:
        '200':
          description: Healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  db:
                    type: string

  /version:
    get:
      summary: Version info
      tags: [System]
      responses:
        '200':
          description: Version
          content:
            application/json:
              schema:
                type: object
                properties:
                  version:
                    type: string

  /api/v1/auth/register:
    post:
      summary: Register new user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, name, password]
              properties:
                email:
                  type: string
                  format: email
                name:
                  type: string
                password:
                  type: string
                  minLength: 8
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Validation error
        '409':
          description: Email already exists

  /api/v1/auth/login:
    post:
      summary: Login
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid credentials

  /api/v1/auth/refresh:
    post:
      summary: Refresh access token
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid refresh token

  /api/v1/auth/me:
    get:
      summary: Get current user
      tags: [Auth]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized

  /api/v1/orgs:
    post:
      summary: Create organization
      tags: [Organizations]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, slug]
              properties:
                name:
                  type: string
                slug:
                  type: string
      responses:
        '201':
          description: Organization created
          content:
            application/json:
              schema:
                type: object
                properties:
                  org:
                    $ref: '#/components/schemas/Org'
                  membership:
                    $ref: '#/components/schemas/Membership'
    get:
      summary: List organizations
      tags: [Organizations]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Organizations list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Org'

  /api/v1/orgs/{orgID}:
    get:
      summary: Get organization
      tags: [Organizations]
      security:
        - bearerAuth: []
      parameters:
        - name: orgID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Organization details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Org'
        '404':
          description: Not found

  /api/v1/orgs/{orgID}/connections:
    post:
      summary: Create connection
      tags: [Connections]
      security:
        - bearerAuth: []
      parameters:
        - name: orgID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, name]
              properties:
                type:
                  type: string
                  enum: [docker, k8s]
                name:
                  type: string
      responses:
        '201':
          description: Connection created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Connection'
    get:
      summary: List connections
      tags: [Connections]
      security:
        - bearerAuth: []
      parameters:
        - name: orgID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Connections list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Connection'

  /api/v1/orgs/{orgID}/connections/{connectionID}/topology:
    get:
      summary: Get topology
      tags: [Topology]
      security:
        - bearerAuth: []
      parameters:
        - name: orgID
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: connectionID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Topology data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Topology'

  /api/v1/orgs/{orgID}/connections/{connectionID}/metrics:
    get:
      summary: Get container metrics
      tags: [Metrics]
      security:
        - bearerAuth: []
      parameters:
        - name: orgID
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: connectionID
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: runtime_id
          in: query
          required: true
          schema:
            type: string
        - name: metric
          in: query
          schema:
            type: string
            enum: [cpu, mem, net_rx, net_tx, disk_r, disk_w]
        - name: start
          in: query
          schema:
            type: string
            format: date-time
        - name: end
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Metrics data

  /api/v1/orgs/{orgID}/connections/{connectionID}/flows:
    get:
      summary: List network flows
      tags: [Flows]
      security:
        - bearerAuth: []
      parameters:
        - name: orgID
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: connectionID
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            default: 1000
      responses:
        '200':
          description: Flows list

  /api/v1/orgs/{orgID}/connections/{connectionID}/vulns:
    get:
      summary: List vulnerability scans
      tags: [Vulnerabilities]
      security:
        - bearerAuth: []
      parameters:
        - name: orgID
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: connectionID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Scans list
    post:
      summary: Record vulnerability scan
      tags: [Vulnerabilities]
      security:
        - bearerAuth: []
      parameters:
        - name: orgID
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: connectionID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '201':
          description: Scan recorded

  /api/v1/orgs/{orgID}/connections/{connectionID}/alerts:
    get:
      summary: List alerts
      tags: [Alerts]
      security:
        - bearerAuth: []
      parameters:
        - name: orgID
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: connectionID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Alerts list

  /api/v1/orgs/{orgID}/connections/{connectionID}/alerts/rules:
    get:
      summary: List alert rules
      tags: [Alerts]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Rules list
    post:
      summary: Create alert rule
      tags: [Alerts]
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Rule created

  /api/v1/orgs/{orgID}/connections/{connectionID}/alerts/channels:
    get:
      summary: List notification channels
      tags: [Alerts]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Channels list
    post:
      summary: Create notification channel
      tags: [Alerts]
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Channel created

  /api/v1/agent/enroll:
    post:
      summary: Agent enrollment
      tags: [System]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enrollment_token]
              properties:
                enrollment_token:
                  type: string
                hostname:
                  type: string
                os:
                  type: string
                kernel:
                  type: string
                cpu_cores:
                  type: integer
                mem_total:
                  type: integer
                agent_version:
                  type: string
      responses:
        '200':
          description: Enrolled successfully

  /api/v1/agent/heartbeat:
    post:
      summary: Agent heartbeat
      tags: [System]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [connection_id]
              properties:
                connection_id:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Heartbeat received

  /api/v1/agent/topology:
    post:
      summary: Sync topology
      tags: [System]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [connection_id]
              properties:
                connection_id:
                  type: string
                  format: uuid
                containers:
                  type: array
                networks:
                  type: array
      responses:
        '200':
          description: Topology synced